refactor shared routines into a library
